backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
jiter==0.14.0
openai==2.32.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
